refactor(resources): fold the flattened list into resources.list instead of a new endpoint
resources.list had no production consumers and its default-collection-only
semantics were dead weight, so repurpose it as the profile-scoped flattened
read (openProcedure, same fail-closed decisions READ gate as its sibling
list endpoints) rather than adding listAcrossCollections alongside it. Adds
access-tier gating cells for the now-open procedure.
