+++ services/api/src/routers/resources/index.ts
@@ -6,7 +6,6 @@ import { createLink } from './createLink';
 import { deleteResourceRouter } from './delete';
 import { detachFromCollection } from './detachFromCollection';
 import { list } from './list';
-import { listAcrossCollections } from './listAcrossCollections';
 import { listByCollection } from './listByCollection';
 import { reorder } from './reorder';
 import { update } from './update';
@@ -15,7 +14,6 @@ import { uploadFile } from './uploadFile';
 export const resourcesRouter = mergeRouters(
   collectionsRouter,
   list,
-  listAcrossCollections,
   listByCollection,
   createLink,
   createDocument,
